graphql client call tests and refactorings
