+++ src/main/java/com/shared/info/service/impl/GreetServiceImpl.java
@@ -1,12 +1,13 @@
-package com.shared.info.service;
+package com.shared.info.service.impl;
 
 import com.shared.info.port.GreetPort;
+import com.shared.info.service.GreetService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-public class GreetServiceImpl implements GreetService{
+public class GreetServiceImpl implements GreetService {
 
     private final GreetPort greetPort;
